+++ server/src/connection/mod.rs
@@ -5,3 +5,5 @@ pub mod io;
 pub mod ping_config;
 pub mod ping_manager;
 pub mod time_manager;
+pub mod channel_tick_buffer_receiver;
+pub mod tick_buffer_receiver;
