--- server/src/tick/mod.rs
@@ -1,3 +1,0 @@
-pub mod channel_tick_buffer_receiver;
-pub mod tick_buffer_receiver;
-pub mod tick_manager;
